requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
torch==1.10.0
transformers==4.15.0
Pillow==8.4.0
gunicorn==20.1.0
